Added try and catch statement to function
